refactor: split vsock-host file operations
